+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 0.4.2
+* [Feature] issue#6 add `setup` method
+```ruby
+class SomeCommand < Act::Command
+  # use setup to pre-inialize some actions.
+  setup do
+    @ins = 1
+  end
+
+  def perform
+    @ins # use @ins
+  end
+end
+```
+
 # 0.4.1
 * fix issue#5